refactor(organizations): extract member status selection
