+++ src/routes/Skills/OrganizationMembersTable.tsx
@@ -3,6 +3,7 @@ import type { BusyAction, MemberView, ProviderGrantView } from "./organization-m
 
 import { MoreHorizontalIcon, PencilIcon, ShieldCheckIcon, Trash2Icon, UserCheckIcon, UserXIcon } from "lucide-react"
 import * as React from "react"
+import { hasMemberStatus, isBulkEditableMember, useMemberStatusSelection } from "./use-member-status-selection.ts"
 import { CachedAvatarImage } from "@/components/CachedAvatarImage"
 import { CopyIconButton } from "@/components/CopyIconButton"
 import { Badge } from "@/components/ui/badge"
@@ -65,77 +66,24 @@ export function MembersTable({
   const { t } = useAppI18n()
   const [removeTarget, setRemoveTarget] = React.useState<MemberView | null>(null)
   const [revokeTarget, setRevokeTarget] = React.useState<ProviderGrantView | null>(null)
-  const [selectedUserIds, setSelectedUserIds] = React.useState<Set<string>>(() => new Set())
   const removeTargetBusy = removeTarget ? busyAction === `remove:${removeTarget.user_id}` : false
   const revokeTargetBusy = revokeTarget ? busyAction === `revokeProviderAccess:${revokeTarget.userId}` : false
-  const showStatusColumn = members.some(hasMemberStatus)
-  const canBulkManage = canManage && showStatusColumn
-  const selectableMembers = React.useMemo(
-    () => (canBulkManage ? members.filter(isBulkEditableMember) : []),
-    [canBulkManage, members],
-  )
-  const selectableUserIds = React.useMemo(
-    () => new Set(selectableMembers.map((member) => member.user_id)),
-    [selectableMembers],
-  )
-  const selectedMembers = React.useMemo(
-    () => selectableMembers.filter((member) => selectedUserIds.has(member.user_id)),
-    [selectableMembers, selectedUserIds],
-  )
-  const selectedEnableUserIds = React.useMemo(
-    () => selectedMembers.filter((member) => member.disable).map((member) => member.user_id),
-    [selectedMembers],
-  )
-  const selectedDisableUserIds = React.useMemo(
-    () => selectedMembers.filter((member) => !member.disable).map((member) => member.user_id),
-    [selectedMembers],
-  )
-  const selectedCount = selectedMembers.length
-  const allSelected = selectableMembers.length > 0 && selectedCount === selectableMembers.length
-  const someSelected = selectedCount > 0 && !allSelected
-  const bulkBusy = busyAction === "enableMembers" || busyAction === "disableMembers"
-
-  React.useEffect(() => {
-    setSelectedUserIds((current) => {
-      const next = new Set([...current].filter((userId) => selectableUserIds.has(userId)))
-      return next.size === current.size ? current : next
-    })
-  }, [selectableUserIds])
-
-  const toggleAll = React.useCallback(
-    (checked: boolean) => {
-      setSelectedUserIds(checked ? selectableUserIds : new Set<string>())
-    },
-    [selectableUserIds],
-  )
-
-  const toggleMember = React.useCallback((userId: string, checked: boolean) => {
-    setSelectedUserIds((current) => {
-      const next = new Set(current)
-      if (checked) {
-        next.add(userId)
-      } else {
-        next.delete(userId)
-      }
-      return next
-    })
-  }, [])
-
-  const enableSelectedMembers = React.useCallback(() => {
-    if (selectedEnableUserIds.length === 0) {
-      return
-    }
-    onEnableMembers(selectedEnableUserIds)
-    setSelectedUserIds(new Set<string>())
-  }, [onEnableMembers, selectedEnableUserIds])
-
-  const disableSelectedMembers = React.useCallback(() => {
-    if (selectedDisableUserIds.length === 0) {
-      return
-    }
-    onDisableMembers(selectedDisableUserIds)
-    setSelectedUserIds(new Set<string>())
-  }, [onDisableMembers, selectedDisableUserIds])
+  const {
+    allSelected,
+    bulkBusy,
+    canBulkManage,
+    disableSelectedMembers,
+    enableSelectedMembers,
+    selectedCount,
+    selectedDisableUserIds,
+    selectedEnableUserIds,
+    selectedUserIds,
+    selectableMembers,
+    showStatusColumn,
+    someSelected,
+    toggleAll,
+    toggleMember,
+  } = useMemberStatusSelection({ busyAction, canManage, members, onDisableMembers, onEnableMembers })
 
   const removeConfirmDialog = (
     <ConfirmDialog
@@ -671,14 +619,6 @@ function MemberStatusCheckbox({
       onChange={(event) => onCheckedChange(event.currentTarget.checked)}
     />
   )
-}
-
-function hasMemberStatus(member: OrganizationMember): member is OrganizationMember & { disable: boolean } {
-  return typeof member.disable === "boolean"
-}
-
-function isBulkEditableMember(member: OrganizationMember): member is OrganizationMember & { disable: boolean } {
-  return member.role !== "creator" && hasMemberStatus(member)
 }
 
 function CopyValueButton({ ariaLabel, copiedLabel, value }: { ariaLabel: string; copiedLabel: string; value: string }) {
